+++ manyun-modules/manyun-admin/src/main/java/com/manyun/admin/domain/query/ConsignmentQuery.java
@@ -28,10 +28,10 @@ public class ConsignmentQuery
     private String buiName;
 
     @ApiModelProperty("0藏品，1盲盒")
-    private Long isType;
+    private Integer isType;
 
     @ApiModelProperty("寄售状态-1=已寄售,2=已锁单(有买方,未支付而已)  3=已售出")
-    private Long consignmentStatus;
+    private Integer consignmentStatus;
 
     @ApiModelProperty("卖家姓名")
     private String sellerName;
