+++ manyun-modules/manyun-admin/src/main/java/com/manyun/admin/domain/vo/CntConsignmentVo.java
@@ -17,10 +17,10 @@ public class CntConsignmentVo
     private String id;
 
     @ApiModelProperty("寄售状态-1=已寄售,2=已锁单(有买方,未支付而已)  3=已售出")
-    private Long consignmentStatus;
+    private Integer consignmentStatus;
 
     @ApiModelProperty("0藏品，1盲盒")
-    private Long isType;
+    private Integer isType;
 
     @ApiModelProperty("收藏家姓名")
     private String collectorName;
